RCMsize: Sample Size Calculation in Reversible Catalytic Models

Sample size and confidence interval calculations in reversible catalytic models, with applications in malaria research. Further details can be found in the paper by Sepúlveda and Drakeley (2015, doi:10.1186/s12936-015-0661-z ).

See also

Author

Maintainer: Márcia Graça marcia_da_graca@hotmail.com [copyright holder]